+++ examples/map_basic.rs
@@ -2,14 +2,12 @@
 
 use bevy::prelude::*;
 use bevy_ecs_tiled::prelude::*;
-use bevy_ecs_tilemap::prelude::*;
 
 fn main() {
     App::new()
         // Bevy default plugins
         .add_plugins(DefaultPlugins)
-        // bevy_ecs_tilemap and bevy_ecs_tiled main plugins
-        .add_plugins(TilemapPlugin)
+        // bevy_ecs_tiled main plugin
         .add_plugins(TiledMapPlugin::default())
         // Add our systems and run the app!
         .add_systems(Startup, startup)