+++ examples/map_spawn_delay.rs
@@ -4,7 +4,6 @@ use std::time::*;
 
 use bevy::prelude::*;
 use bevy_ecs_tiled::prelude::*;
-use bevy_ecs_tilemap::prelude::*;
 
 mod helper;
 
@@ -14,8 +13,7 @@ fn main() {
         .add_plugins(DefaultPlugins)
         // Examples helper plugin (does not matter for this example)
         .add_plugins(helper::HelperPlugin)
-        // bevy_ecs_tilemap and bevy_ecs_tiled main plugins
-        .add_plugins(TilemapPlugin)
+        // bevy_ecs_tiled main plugin
         .add_plugins(TiledMapPlugin::default())
         // Add our systems and run the app!
         .add_systems(Startup, startup)