+++ examples/orientation_isometric.rs
@@ -12,8 +12,7 @@ fn main() {
         .add_plugins(DefaultPlugins)
         // Examples helper plugins: for this example, contains the logic to switch between maps
         .add_plugins(helper::HelperPlugin)
-        // bevy_ecs_tilemap and bevy_ecs_tiled main plugins
-        .add_plugins(TilemapPlugin)
+        // bevy_ecs_tiled main plugin
         .add_plugins(TiledMapPlugin::default())
         // Enable debug informations
         .add_plugins(TiledDebugPluginGroup)