+++ examples/physics_custom.rs
@@ -6,7 +6,6 @@ use bevy::{
     prelude::*,
 };
 use bevy_ecs_tiled::prelude::*;
-use bevy_ecs_tilemap::prelude::*;
 
 mod helper;
 
@@ -16,8 +15,7 @@ fn main() {
         .add_plugins(DefaultPlugins)
         // Examples helper plugin (does not matter for this example)
         .add_plugins(helper::HelperPlugin)
-        // bevy_ecs_tilemap and bevy_ecs_tiled main plugins
-        .add_plugins(TilemapPlugin)
+        // bevy_ecs_tiled main plugin
         .add_plugins(TiledMapPlugin::default())
         // bevy_ecs_tiled physics plugin: this is where we select which physics backend to use
         .add_plugins(TiledPhysicsPlugin::<MyCustomPhysicsBackend>::default())